[BE] feat#180: 주소 변환 API 기능 컨트롤러 구현
:tada: :tada: :tada:
이현수 <znight1020@naver.com>님이 Closes #180을 완료했습니다. 

커밋 목록
* feat: 주소 변환 기능 컨트롤러 구현

* refactor: 재발급 헤더가 있는 경우만 로그아웃 진행

승인자
Approved-by: 김도예 <kmdy125@gmail.com>

+++ module-api/src/main/java/com/e205/auth/jwt/handler/JwtRequestProcessHandler.java
@@ -46,10 +46,12 @@ public class JwtRequestProcessHandler {
   @DeleteMapping("/auth/logout")
   public void logout(HttpServletRequest request) {
     String headerValue = request.getHeader(REFRESH_TOKEN_HEADER_NAME);
-    String refreshToken = extractRefreshToken(headerValue);
-    Integer memberId = jwtProvider.getMemberId(refreshToken);
-    if (jwtRepository.existByMemberId(memberId)) {
-      jwtRepository.deleteRefreshTokenByMemberId(memberId);
+    if(headerValue != null) {
+      String refreshToken = extractRefreshToken(headerValue);
+      Integer memberId = jwtProvider.getMemberId(refreshToken);
+      if (jwtRepository.existByMemberId(memberId)) {
+        jwtRepository.deleteRefreshTokenByMemberId(memberId);
+      }
     }
   }
